+++ tuning/acceleration/plugins/__init__.py
@@ -3,6 +3,6 @@ from .framework_plugin import AccelerationPlugin, get_relevant_configuration_sec
 # can this be automated?
 from .framework_plugin_autogptq import AutoGPTQAccelerationPlugin
 from .framework_plugin_bnb import BNBAccelerationPlugin
-from .framework_plugin_unsloth_autogptq import UnslothAutoGPTQAccelerationPlugin
-from .framework_plugin_unsloth import UnslothStackableAccelerationPlugin
+from .framework_plugin_unsloth_standalone import UnslothAutoGPTQAccelerationPlugin
+from .framework_plugin_unsloth_stackable import UnslothStackableAccelerationPlugin
